Addressed more review feedback and fixed a bug in Actor
